Fix for Bug #1093823
The mass of the tau lepton was not correctly taken into account in the
phase-space set-up: for j_fks initial state, s-hat could be smaller
than the sum of all the final state masses, because the tau lepton
mass was ignored, triggering a safety check in the code that let the
code to stop running. The fix makes sure that the tau lepton mass is
taken into account, as well as making the safety check a bit more
stringent to let it also apply if the resonance was not a
Breit-Wigner.