Fixed several points raised by olivier (the only one not fixed is the jpg output) sum_html.f  has been removed from the NLO source directory
FIX 2) The page index.html doesn't have the link pointing through crossx.html (really annoying) FIX

FIX 3) They are no update status for when the shower run FIX

(I suspect the same for MadSpin) 

FIX 4) When the code is finished, the status bar (information at the top) is still present (it should be removed). FIX

FIX 5) The file output in the HTML seems to be present before the shower. So this should be an unphysical event sample. We should (at least) also put a link to the events file produced by the shower.

FIX 6) I've run in NLO mode and the crossx.html didn't show up automatically. Is this intended? FIX

FIX, removed button7) When events are present, a button launch detector simulation is present...
but when I click on it I have the following message:
Launch ./bin/aMCatNLO in a shell, and run the following command: pythia run_01
So those two are not coherent. (The command is wrong and didn't correspond to the button)

FIX 8) Shouldn't we include in the html the mode of aMC@NLO used for the computation (LO/NLO and Fixed order or not) This is really usefull information which are not really present in any file.

FIX9)when an error occur in the code, the html page is not updated to inform of that crash report. (and continue to indicate that the code is running)