Fixed a small bug of the previous revision.
MARCO: this version works, with the inconvenience that I write everything in
MCatNLO/shower_card_set.dat and then, if this file is there, I mv it to
MCatNLO/shower_card.dat.
This works in general BUT for example if you are using herwig6, employ the
'set' command for the shower and then you do 'set parton_shower pythia6q',
the file shower_card_set.dat written in MCatNLO is for tailored for hewrig6,
and thus the shower crashes.
The proper solution is to write the Cards/shower_card after each 'set' command,
so that the translator to the MCatNLO/shower_card.dat will always work.

Relevant tihings to modify are at least in common_run_interface.py (around
line 2372) and in amcatnlo_run_interface.py (arount line 2265).