1. Big improvement in the fortran writeout of NLO models for which some CT couplings are expressed in terms of CT parameters. The resulting expression is no longer fully expanded, but instead the CT parameters are defined as regular ones for the finite part and poles and they are then reused in the expression of the CT couplings. This yields an output much more efficient, both at compile and run time.
Notice that when using CTparameters in the value of a coupling, one is only allowed to use:
  > Operators +,-,/,* and the operation complexconjugate
  > *exactly one* (not less, not more) CTParam per term added or subtracted.

(if the CT coupling only has a finite part it is of course possible to use no CT params at all).  
If more complicated functions are needed, they must either be put in the definition of a CT parameter or the use of CTparameter should be avoided all together in this case and the CT coupling value should be defined as a dictionary directly.

2. Added and fixed the loop_qcd_qed_sm_Gmu__CMS__ model which used the 'a==b==0.0' syntax and 'math.pi' instead of 'cmath.pi' which are both forbidden.

3. Must check and fix tests now.